+++ hutool-extra/src/main/java/cn/hutool/extra/servlet/JakartaServletUtil.java
@@ -5,6 +5,7 @@ import cn.hutool.core.bean.copier.CopyOptions;
 import cn.hutool.core.bean.copier.ValueProvider;
 import cn.hutool.core.collection.ArrayIter;
 import cn.hutool.core.collection.IterUtil;
+import cn.hutool.core.collection.ListUtil;
 import cn.hutool.core.exceptions.UtilException;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IORuntimeException;
@@ -20,12 +21,12 @@ import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.ReflectUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.core.util.URLUtil;
-
 import jakarta.servlet.ServletOutputStream;
 import jakarta.servlet.ServletRequest;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
 import java.io.File;
@@ -35,11 +36,14 @@ import java.io.PrintWriter;
 import java.io.Writer;
 import java.lang.reflect.Type;
 import java.nio.charset.Charset;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
 import java.util.Enumeration;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.List;
 
 /**
  * Servlet相关工具类封装
@@ -294,6 +298,43 @@ public class JakartaServletUtil {
 		while (names.hasMoreElements()) {
 			name = names.nextElement();
 			headerMap.put(name, request.getHeader(name));
+		}
+
+		return headerMap;
+	}
+
+	/**
+	 * 获取请求所有的头（header）信息
+	 *
+	 * @param request 请求对象{@link HttpServletRequest}
+	 * @return header值
+	 * @since 6.0.0
+	 */
+	public static Map<String, List<String>> getHeadersMap(final HttpServletRequest request) {
+		final Map<String, List<String>> headerMap = new LinkedHashMap<>();
+
+		final Enumeration<String> names = request.getHeaderNames();
+		String name;
+		while (names.hasMoreElements()) {
+			name = names.nextElement();
+			headerMap.put(name, ListUtil.list(false, request.getHeaders(name)));
+		}
+
+		return headerMap;
+	}
+
+	/**
+	 * 获取响应所有的头（header）信息
+	 *
+	 * @param response 响应对象{@link HttpServletResponse}
+	 * @return header值
+	 */
+	public static Map<String, Collection<String>> getHeadersMap(final HttpServletResponse response) {
+		final Map<String, Collection<String>> headerMap = new HashMap<>();
+
+		final Collection<String> names = response.getHeaderNames();
+		for (final String name : names) {
+			headerMap.put(name, response.getHeaders(name));
 		}
 
 		return headerMap;